remove redundant isStandardInputScript
